+++ src/features/CreateCategoryView/components/ArrowButtons/index.tsx
