+++ src/features/HomePageView/components/AdvantageCard/styles.module.scss
@@ -1,0 +1,22 @@
+.advantageCard {
+    display: flex;
+    gap: 24px;
+    padding: 24px;
+    width: 40%;
+    border-radius: 24px;
+    background: var(--light-color);
+    box-shadow: 0px 1px 1px 0px rgba(0, 0, 0, 0.10), 0px 2px 2px 0px rgba(0, 0, 0, 0.09), 0px 5px 3px 0px rgba(0, 0, 0, 0.05), 0px 8px 3px 0px rgba(0, 0, 0, 0.01), 0px 13px 4px 0px rgba(0, 0, 0, 0.00);
+
+    &_title {
+        margin: 0;
+        font-size: 16px;
+        font-weight: 600;
+    }
+
+    &_desc {
+        font-size: 16px;
+        font-weight: 400;
+        margin: 0;
+        margin-top: 8px;
+    }
+}
