--- src/features/HomePageView/components/guide-block/guideBlock.module.scss
@@ -1,68 +1,0 @@
-.guideBlock {
-    display: flex;
-    align-items: center;
-    gap: 24px;
-    margin-block: 24px;
-    background-color: var(--light-color);
-    border-radius: 24px;
-    box-shadow: 0px 1px 3px 0px rgba(0, 0, 0, 0.10), 0px 5px 5px 0px rgba(0, 0, 0, 0.09), 0px 12px 7px 0px rgba(0, 0, 0, 0.05), 0px 21px 9px 0px rgba(0, 0, 0, 0.01), 0px 33px 9px 0px rgba(0, 0, 0, 0.00);
-    padding: 24px;
-
-    &_info {
-        flex: 1;
-        padding: 24px;
-    }
-
-    &_title {
-        font-size: 16px;
-        font-weight: 600;
-        margin: 0;
-    }
-
-    &_list {
-        margin: 0;
-        margin-top: 24px;
-        font-size: 16px;
-        display: flex;
-        flex-direction: column;
-        gap: 24px;
-    }
-}
-
-.block {
-    background-color: var(--accent-color);
-    border-radius: 8px;
-    flex: 2;
-    height: 500px;
-}
-
-
-@media (max-width: 1200px) {
-    .guideBlock {
-       flex-direction: column !important;
-    }
-    .block {
-        width: 100%;
-        flex: none;
-    }
-}
-
-@media (max-width: 600px) {
-    .guideBlock {
-
-        width: 88%;
-        margin-inline:auto;
-        &_info {
-            padding: 0;
-        }
-        &_list {
-            gap: 20px;
-            margin-top: 20px;
-            padding: 0;
-            list-style-position: inside;
-        }
-    }
-    .block {
-        height: 300px;
-    }
-}
