+++ src/features/HomePageView/containers/Guide/styles.module.scss
@@ -1,0 +1,11 @@
+.guide {
+    background-color: var(--green-color);
+    padding-block: 64px;
+    color: var(--dark-color);
+
+    &_title {
+        font-size: 32px;
+        font-weight: 600;
+        margin: 0;
+    }
+}
