--- src/features/HomePageView/containers/guide/guide.module.scss
@@ -1,11 +1,0 @@
-.guide {
-    background-color: var(--green-color);
-    padding-block: 64px;
-    color: var(--dark-color);
-
-    &_title {
-        font-size: 32px;
-        font-weight: 600;
-        margin: 0;
-    }
-}
