+++ src/features/HomePageView/containers/guide/index.tsx
@@ -1,19 +1,19 @@
-import { FC } from 'react'
-import styles from './guide.module.scss'
-import CarouselC from '../../components/carousel/Carousel';
-import ContainerL from '../../../../components/Container/Container';
+import { FC } from "react";
+import styles from "./styles.module.scss";
+import CarouselC from "../../components/Carousel";
+import ContainerL from "../../../../components/Container/Container";
 
 const GuideSection: FC = () => {
-  return (
-    <section>
-      <div className={styles.guide}>
-        <ContainerL>
-          <h2 className={styles.guide_title}>Гайд для користувачів для заповнення інтернет-довідок і заяв:</h2>
-        </ContainerL>
-        <CarouselC />
-      </div>
-    </section>
-  )
-}
+    return (
+        <section>
+            <div className={styles.guide}>
+                <ContainerL>
+                    <h2 className={styles.guide_title}>Гайд для користувачів для заповнення інтернет-довідок і заяв:</h2>
+                </ContainerL>
+                <CarouselC />
+            </div>
+        </section>
+    );
+};
 
-export default GuideSection
+export default GuideSection;
