+++ src/features/RequestProcessingView/components/MessageDocument/index.tsx
@@ -1,0 +1,28 @@
+import { FC } from "react";
+import { Button } from "@mantine/core";
+import { Icon } from "@iconify/react";
+
+
+interface MessageDocumentProps {
+    label: string;
+    documentLink: string;
+}
+
+
+const MessageDocument: FC<MessageDocumentProps> = ({ label, documentLink }) => {
+    return (
+        <a href={documentLink}>
+            <Button
+                w={217}
+                radius="xl"
+                variant="outline"
+                color="#02808F"
+                leftSection={<Icon width={24} height={24} icon="basil:document-outline" />}
+            >
+                {label}
+            </Button>
+        </a>
+    )
+}
+
+export default MessageDocument;
